content-generation/pyproject.toml
[project]
name = "agentic-learning-content-generation"
version = "0.1.0"
description = "Content generation pipeline for the agentic learning app"
requires-python = "==3.13.*"
dependencies = [
    "google-genai>=1.0.0",
    "google-auth>=2.0.0",
    "google-cloud-aiplatform>=1.60.0",
    "pyyaml>=6.0",
    "pydantic-settings>=2.0",
    "tiktoken>=0.12.0",
]

[project.optional-dependencies]
# GCS backend — required when GCS_PIPELINE_BUCKET is set (i.e. on Cloud Run).
# Not needed for local development.
gcs = [
    "google-cloud-storage>=2.0",
]
# Extra deps for seed_db.py (Cloud SQL connector + pg8000 driver).
# Not needed for generate_content.py or embed_content.py.
seed = [
    "cloud-sql-python-connector[pg8000]>=1.7.0",
    "pg8000>=1.31.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "ruff>=0.15.7",
]

[tool.setuptools]
py-modules = [
    "config",
    "storage",
    "review_models",
    "token_usage_log",
    "generate_content",
    "embed_content",
    "seed_db",
    "validate_embeddings",
    "validate_db",
    "count_tokens",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
